HIP 15 | David | Refactoring patient controller test by extracting Test server for reuse by other integration tests
