+++ test/In.ProjectEKA.HipServiceTest/Common/TestServer/AuthenticatedTestRequestMiddleware.cs
@@ -1,0 +1,26 @@
+﻿namespace In.ProjectEKA.HipServiceTest.Common.TestServer
+{
+    using System.Security.Claims;
+    using System.Threading.Tasks;
+    using Microsoft.AspNetCore.Http;
+
+    public class AuthenticatedTestRequestMiddleware
+    {
+        private readonly RequestDelegate _next;
+
+        public AuthenticatedTestRequestMiddleware(RequestDelegate next)
+        {
+            _next = next;
+        }
+
+        public async Task Invoke(HttpContext context)
+        {
+            var claims = new[] { new Claim(ClaimTypes.Name, "Test user") };
+            var identity = new ClaimsIdentity(claims, "Test");
+            var principal = new ClaimsPrincipal(identity);
+            context.User = principal;
+
+            await _next(context);
+        }
+    }
+}
