+++ test/In.ProjectEKA.HipServiceTest/In.ProjectEKA.HipServiceTest.csproj
@@ -36,4 +36,7 @@
         <ProjectReference Include="..\..\src\In.ProjectEKA.HipLibrary\In.ProjectEKA.HipLibrary.csproj" />
         <ProjectReference Include="..\..\src\In.ProjectEKA.HipService\In.ProjectEKA.HipService.csproj" />
     </ItemGroup>
+    <ItemGroup>
+      <Folder Include="Common\TestServer\" />
+    </ItemGroup>
 </Project>
